Allow fetchers to export JSON files directly
Instead of having each fetcher return a CSV string, this update allows
each fetcher to create it's own data files directly. This should make
things more open for extension in the future, because a fetcher doesn't
need to be tied to a specific return value. Instead, it can format files
directly however it likes (or append to multiple files if needed).

Co-authored-by: Andrew Henry <ajhenry@github.com>