Remove graphql workflow
Because the build step now includes the data files, we don't need a
separate action to do those. We can instead remove this and depend on
the nextjs workflow to automatically load the data it needs.

If we need CSV functionality, we can either 1) write an action/workflow
to download a CSV for you or 2) add a CSV download button to the UI.

Co-authored-by: Andrew Henry <ajhenry@github.com>