Render basic repositories table
This starts to actually incorporate the data we generate into the frontend
by creating a basic repositories table to replace the placeholder "PerformanceHistoryTable"
that was already there. Right away, this lets you:

- View all of the public repositories
- Filter by license type
- Filter by repository name

We also update the `DashboardExample` to read the name of the Organization
from the data file to make the Organization Name a bit more dynamic.

We'll need to figure out how to do something for the logo, but going to hold
off on that for now.