Use Primer component for RepositorySelector
There seemed to be a bug with the Tremor selector where clearing the search
filter by closing the popover didn't restore the filtered list of items.

This commit switches to using the Primer `SelectPanel` instead. I added a
`RepositorySelector` component to manage the state for this. There may have
been a simpler fix - I'll probably take a look into the Tremor component to
see if I'm missing something.

We should decide what Component library we want to use for this - it might
more sense to roll with Tremor if we're going to use that for the charting.

I can also look into opening a PR upstream if there is in fact a bug in the
Select component.

Finally, I added some logic to make the Primer theme match the Tremor theme.
This is a downside of using two different component libraries.