Add first pass at TS Backend
We're exploring using TypeScript and octokit for our backend instead of
Go. This adds a basic typescript structure and command to fetch data.

The basic idea is that we can now use pipeline functions to append data
to our result. This should make it easier to compose data. If this continues
to look promising, we can add a test to compare to the results of the Go backend
to ensure we're getting the same data.

Co-authored-by: Andrew Henry <ajhenry@github.com>