Feature: Add topics to the UI
- Adds the new topic header as a comma separated list to the table, along
with a simple popover to show the full results. Also added sorting and filtering
similar to licenses.
- Updated the date display to supress a warning about server side rendering
since the client browser may be different from the server in local dev.
- Replaced the `Repo` type with `RepositoryResult` from our TS library.
We previously were creating the type based on the first object result, but this
broken down because `topics` is potentially an empty array, so we can't infer that it
will have a value